requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
flask-cors